+++ test/src/test_stream_reader.cpp
@@ -42,16 +42,16 @@ TEST(test_stream_reader, read_bytes)
     uint8_t byte8 = 0;
     uint8_t byte9 = 0;
 
-    reader.read<endian::u8>(byte0);
-    reader.read<endian::u8>(byte1);
-    reader.read<endian::u8>(byte2);
-    reader.read<endian::u8>(byte3);
-    reader.read<endian::u8>(byte4);
-    reader.read<endian::u8>(byte5);
-    reader.read<endian::u8>(byte6);
-    reader.read<endian::u8>(byte7);
-    reader.read<endian::u8>(byte8);
-    reader.read<endian::u8>(byte9);
+    reader.read_bytes<1>(byte0);
+    reader.read_bytes<1>(byte1);
+    reader.read_bytes<1>(byte2);
+    reader.read_bytes<1>(byte3);
+    reader.read_bytes<1>(byte4);
+    reader.read_bytes<1>(byte5);
+    reader.read_bytes<1>(byte6);
+    reader.read_bytes<1>(byte7);
+    reader.read_bytes<1>(byte8);
+    reader.read_bytes<1>(byte9);
 
     EXPECT_EQ(0U, byte0);
     EXPECT_EQ(1U, byte1);
@@ -70,19 +70,49 @@ TEST(test_stream_reader, read_bytes)
     uint8_t byte11 = initial_value;
     uint8_t byte12 = initial_value;
     uint8_t byte13 = initial_value;
-    reader.read<endian::u8>(byte10);
+    reader.read_bytes<1>(byte10);
     EXPECT_TRUE(bool(error));
-    reader.read<endian::u8>(byte11);
+    reader.read_bytes<1>(byte11);
     EXPECT_TRUE(bool(error));
-    reader.read<endian::u8>(byte12);
+    reader.read_bytes<1>(byte12);
     EXPECT_TRUE(bool(error));
-    reader.read<endian::u8>(byte13);
+    reader.read_bytes<1>(byte13);
     EXPECT_TRUE(bool(error));
 
     EXPECT_EQ(initial_value, byte10);
     EXPECT_EQ(initial_value, byte11);
     EXPECT_EQ(initial_value, byte12);
     EXPECT_EQ(initial_value, byte13);
+}
+
+TEST(test_stream_reader, peek_bytes)
+{
+    std::vector<uint8_t> buffer {0, 1, 2};
+    std::error_code error;
+    bnb::stream_reader<endian::big_endian> reader(
+        buffer.data(), buffer.size(), error);
+
+    uint8_t byte0 = 0;
+    uint8_t byte1 = 0;
+    uint8_t byte2 = 0;
+
+    reader.peek_bytes<1>(byte0, 0);
+    reader.peek_bytes<1>(byte1, 1);
+    reader.peek_bytes<1>(byte2, 2);
+
+    EXPECT_EQ(0U, byte0);
+    EXPECT_EQ(1U, byte1);
+    EXPECT_EQ(2U, byte2);
+    EXPECT_TRUE(!error);
+
+    // check error handling
+
+    uint8_t byte3 = 0;
+    reader.peek_bytes<1>(byte3, 3);
+    EXPECT_FALSE(!error);
+    uint8_t byte4 = 0;
+    reader.peek_bytes<1>(byte4, 4);
+    EXPECT_FALSE(!error);
 }
 
 TEST(test_stream_reader, skip)
@@ -104,16 +134,16 @@ TEST(test_stream_reader, skip)
     uint8_t byte9 = 0;
 
     auto skipped = reader.skip(5);
-    skipped.read<endian::u8>(byte0);
-    skipped.read<endian::u8>(byte1);
-    skipped.read<endian::u8>(byte2);
-    skipped.read<endian::u8>(byte3);
-    skipped.read<endian::u8>(byte4);
-    reader.read<endian::u8>(byte5);
-    reader.read<endian::u8>(byte6);
-    reader.read<endian::u8>(byte7);
-    reader.read<endian::u8>(byte8);
-    reader.read<endian::u8>(byte9);
+    skipped.read_bytes<1>(byte0);
+    skipped.read_bytes<1>(byte1);
+    skipped.read_bytes<1>(byte2);
+    skipped.read_bytes<1>(byte3);
+    skipped.read_bytes<1>(byte4);
+    reader.read_bytes<1>(byte5);
+    reader.read_bytes<1>(byte6);
+    reader.read_bytes<1>(byte7);
+    reader.read_bytes<1>(byte8);
+    reader.read_bytes<1>(byte9);
 
     EXPECT_EQ(0U, byte0);
     EXPECT_EQ(1U, byte1);
@@ -140,7 +170,11 @@ TEST(test_stream_reader, skip)
 
 TEST(test_stream_reader, seek)
 {
-    std::vector<uint8_t> buffer = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+    std::vector<uint8_t> buffer =
+        {
+            0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 255, 0, 0, 0, 0, 0, 0, 2
+        };
+
     std::error_code error;
     bnb::stream_reader<endian::big_endian> reader(
         buffer.data(), buffer.size(), error);
@@ -150,29 +184,36 @@ TEST(test_stream_reader, seek)
     uint8_t byte7 = 0;
     uint8_t byte8 = 0;
     uint8_t byte9 = 0;
+    int64_t byte10 = 0;
 
     reader.seek(10);
     reader.seek(5);
-    reader.read<endian::u8>(byte5);
-    reader.read<endian::u8>(byte6);
-    reader.read<endian::u8>(byte7);
-    reader.read<endian::u8>(byte8);
-    reader.read<endian::u8>(byte9);
+    reader.read_bytes<1>(byte5);
+    reader.read_bytes<1>(byte6);
+    reader.read_bytes<1>(byte7);
+    reader.read_bytes<1>(byte8);
+    reader.read_bytes<1>(byte9);
+    reader.peek_bytes<8>().expect_eq(-72057594037927934);
+    reader.peek_bytes<8>().expect_eq(18374686479671623682U);
+    reader.read_bytes<8>(byte10).expect_eq(-72057594037927934);
 
     EXPECT_EQ(5U, byte5);
     EXPECT_EQ(6U, byte6);
     EXPECT_EQ(7U, byte7);
     EXPECT_EQ(8U, byte8);
     EXPECT_EQ(9U, byte9);
+    EXPECT_EQ(-72057594037927934, byte10);
 
-    reader.seek(11);
+    EXPECT_TRUE(!error);
 
-    EXPECT_TRUE((bool)error);
+    reader.seek(reader.size() + 1);
+
+    EXPECT_FALSE(!error);
 
     // Make sure we stay in error state even though we read back to a
     // valid point
     reader.seek(1);
-    EXPECT_TRUE((bool)error);
+    EXPECT_FALSE(!error);
 }
 
 TEST(test_stream_reader, read_bits)
@@ -185,7 +226,7 @@ TEST(test_stream_reader, read_bits)
     bool first_field = false;
     uint8_t third_field = 0;
 
-    reader.read_bits<endian::u8, bitter::msb0, 1, 5, 2>()
+    reader.read_bits<bitter::u8, bitter::msb0, 1, 5, 2>()
     .get<0>(first_field)
     .get<1>().expect_eq(4)
     .get<2>(third_field);
@@ -198,13 +239,13 @@ TEST(test_stream_reader, read_bits)
     uint8_t some_field = 42;
 
     // check that if we read too much data we will get an error
-    reader.read_bits<endian::u8, bitter::msb0, 8>() // force error
+    reader.read_bits<bitter::u8, bitter::msb0, 8>() // force error
     .get<0>(some_field);
     EXPECT_TRUE((bool) error);
     EXPECT_EQ(42U, some_field);
 
     // Check that we can still read bits without crashing
-    reader.read_bits<endian::u8, bitter::msb0, 8>()
+    reader.read_bits<bitter::u8, bitter::msb0, 8>()
     .get<0>(some_field);
 
     EXPECT_TRUE((bool) error);